[Decomp][Player] Cohesion split: extract zone/area update methods to PlayerZone.cpp
